refactor: change api hooks structure
