+++ src/models/Posts.ts
@@ -1,0 +1,6 @@
+export type PostType = {
+  userId: number;
+  id: number;
+  title: string;
+  body: string;
+};
