+++ src/services/posts/index.ts
@@ -1,0 +1,5 @@
+export { useCreatePost } from "./useCreatePost";
+export { useUpdatePost } from "./useUpdatePost";
+export { useDeletePost } from "./useDeletePost";
+export { useFetchPost } from "./useFetchPost";
+export { useFetchPosts } from "./useFetchPosts";
